@@ -17,30 +17,31 @@
     "build:prod": "npm run clean; npm i; npm run build:prod-app && npm run build:server",
     "build:app": "npm run prebuild:app && ng build && ng build --app mte-client-universal",
     "build:staging-app": "npm i; npm run prebuild:app && ng build --prod --environment=staging && ng build --app mte-client-universal --prod --environment=staging",
-    "build:prod-app": "npm run prebuild:app && ng build --prod --environment=prod && ng build --prod --no-aot --environment=prod --app mte-client-universal",
+    "build:prod-app": "npm run prebuild:app && ng build --prod --environment=prod && ng build --prod --aot=false --environment=prod --app mte-client-universal",
     "build:server": "npm run prebuild:api && webpack --config mte-api/config/webpack.server-config.js",
     "test:app": "ng test",
     "lint:app": "ng lint",
     "e2e:app": "ng e2e",
     "watch:server": "webpack --config mte-api/config/webpack.server-config.js --watch",
+    "start:local": "node dist/server",
     "start": "node dist/server",
     "test": "karma start ./karma.conf.js",
     "pree2e": "webdriver-manager update --standalone false --gecko false --quiet",
     "e2e": "protractor ./mte-client/protractor.conf.js"
   },
   "dependencies": {
-    "@angular/animations": "5.1.2",
-    "@angular/common": "5.1.2",
-    "@angular/compiler": "5.1.2",
-    "@angular/compiler-cli": "5.1.2",
-    "@angular/core": "5.1.2",
-    "@angular/forms": "5.1.2",
-    "@angular/http": "5.1.2",
-    "@angular/language-service": "5.0.3",
-    "@angular/platform-browser": "5.1.2",
-    "@angular/platform-browser-dynamic": "5.1.2",
-    "@angular/platform-server": "5.1.2",
-    "@angular/router": "5.1.2",
+    "@angular/animations": "6.0.2",
+    "@angular/common": "6.0.2",
+    "@angular/compiler": "6.0.2",
+    "@angular/compiler-cli": "6.0.2",
+    "@angular/core": "6.0.2",
+    "@angular/forms": "6.0.2",
+    "@angular/http": "6.0.2",
+    "@angular/language-service": "6.0.2",
+    "@angular/platform-browser": "6.0.2",
+    "@angular/platform-browser-dynamic": "6.0.2",
+    "@angular/platform-server": "6.0.2",
+    "@angular/router": "6.0.2",
     "@easypost/api": "3.1.1",
     "@ngrx/effects": "5.0.0",
     "@ngrx/store": "5.0.0",
@@ -156,7 +160,7 @@
     "ts-loader": "2.3.1",
     "ts-node": "3.2.0",
     "tslint": "5.3.2",
-    "typescript": "2.4.2",
+    "typescript": "2.7.2",
     "url-loader": "0.5.7"
   }
 }
