@@ -6,7 +6,7 @@ import {
     Router,
     RouterStateSnapshot,
 } from '@angular/router'
-import { map } from 'rxjs/operators/map'
+import { map } from 'rxjs/operators'
 
 import { UserRole } from '@mte/common/models/enums/user-role'
 import { AppRoutes } from '../../constants/app-routes'
