@@ -1,8 +1,6 @@
 import { HttpClient } from '@angular/common/http'
 import { Injectable } from '@angular/core'
-import { Observable } from 'rxjs/Observable'
-import { ReplaySubject } from 'rxjs/ReplaySubject'
-import { Subject } from 'rxjs/Subject'
+import { Observable, ReplaySubject, Subject } from 'rxjs'
 
 import { ApiEndpoints } from '@mte/common/constants/api-endpoints'
 import { SimpleError } from '@mte/common/lib/ng-modules/http'
