@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core'
 import { NavigationStart, Router } from '@angular/router'
-import { Observable } from 'rxjs/Observable'
+import { Observable } from 'rxjs'
 import { filter } from 'rxjs/operators'
 
 @Injectable()
