@@ -1,8 +1,6 @@
 import { Injectable } from '@angular/core'
-import { Observable } from 'rxjs/Observable'
-import { ReplaySubject } from 'rxjs/ReplaySubject'
-import { Subject } from 'rxjs/Subject'
-import { delay } from 'rxjs/operators/delay'
+import { Observable, ReplaySubject, Subject } from 'rxjs'
+import { delay } from 'rxjs/operators'
 
 @Injectable()
 export class UtilService {
