@@ -8,8 +8,7 @@ import { TaxonomyTerm } from '@mte/common/models/api-models/taxonomy-term'
 import { GetProductsFilter, GetProductsFilterType, GetProductsRequest } from '@mte/common/models/api-requests/get-products.request'
 import { BootstrapBreakpointKey } from '@mte/common/models/enums/bootstrap-breakpoint-key'
 import { cloneDeep } from 'lodash'
-import { BehaviorSubject } from 'rxjs/BehaviorSubject'
-import { Observable } from 'rxjs/Observable'
+import { BehaviorSubject, Observable } from 'rxjs'
 import { map, takeWhile } from 'rxjs/operators'
 import { ShopRouterLinks } from '../../constants/shop-router-links'
 import { ProductService } from '../../services'
