@@ -1,6 +1,6 @@
 import { HttpClient } from '@angular/common/http'
 import { Injectable } from '@angular/core'
-import { Observable } from 'rxjs/Observable'
+import { Observable } from 'rxjs'
 
 import { AppConfig } from '@mte/app-config'
 import { ApiEndpoints } from '@mte/common/constants/api-endpoints'
