@@ -5,7 +5,7 @@ import {
     HttpRequest,
 } from '@angular/common/http'
 import { Injectable } from '@angular/core'
-import { Observable } from 'rxjs/Observable'
+import { Observable } from 'rxjs'
 
 import { AppConfig } from '../../../../app-config'
 
