@@ -1,6 +1,5 @@
-import { Observable } from 'rxjs/Observable'
-import { ReplaySubject } from 'rxjs/ReplaySubject'
-import { delay } from 'rxjs/operators/delay'
+import { Observable, ReplaySubject } from 'rxjs'
+import { delay } from 'rxjs/operators'
 
 export function timeout(timeout: number): Observable<any > {
     const subject = new ReplaySubject(1)
