@@ -1,14 +1,18 @@
+require('dotenv').config()
+
 const webpack = require('webpack')
 const path = require('path')
 const fs = require('fs')
 const CircularDependencyPlugin = require('circular-dependency-plugin')
-require('dotenv').config()
 const packageJson = require('../../package.json')
 const version = packageJson.version
 
 // Exclude all packages from the build, since your node server
 // has access to node_modules directly.
-const externals = {}
+const externals = {
+	'dotenv': 'commonjs dotenv'
+}
+
 fs.readdirSync(path.resolve(__dirname, '../../node_modules'))
 	.filter(x => ['.bin'].indexOf(x) === -1)
 	.forEach(mod => {
@@ -28,6 +32,7 @@ fs.readdirSync(path.resolve(__dirname, '../../node_modules'))
 module.exports = {
 	name: 'server',
 	target: 'node',
+	mode: process.env.ENVIRONMENT === 'development' ? 'development' : 'production',
 	externals,
 	devtool: 'source-map',
 	entry: path.resolve(__dirname, '../server.ts'),
