@@ -1,6 +1,5 @@
 import { Injectable } from '@angular/core'
-import { ReplaySubject } from 'rxjs/ReplaySubject'
-
+import { ReplaySubject } from 'rxjs'
 import { SimpleError } from './http.models'
 
 @Injectable()
