@@ -1,3 +1,4 @@
+import { Price } from '@mte/common/models/api-models/price'
 import { arrayProp, prop, schema, MongooseDocument, Ref } from '../../lib/goosetype'
 import { CartProduct } from '../interfaces/ui/cart-product'
 import { Discount } from './discount'
@@ -5,10 +6,9 @@ import { Product } from './product'
 
 @schema(Cart)
 export class Cart extends MongooseDocument {
-    public displayItems?: CartProduct[]
     @prop() public count?: number
     @arrayProp({ itemsRef: Product }) public items: Ref<Product>[]
-    @prop() public subTotal: number
-    @prop() public total: number
+    @prop() public subTotal: Price
+    @prop() public total: Price
     @arrayProp({ itemsRef: Discount }) public discounts?: Ref<Discount>[]
 }
