@@ -4,5 +4,5 @@ import { Currency } from '../enums/currency'
 @schema(Price, { _id: false })
 export class Price extends MongooseDocument {
     @prop() public amount: number
-    @prop({ enum: Currency }) public currency: string
+    @prop({ enum: Currency, type: String }) public currency: Currency
 }
