@@ -1,4 +1,6 @@
 import { Document } from 'mongoose'
+import { Discount } from '../../api-models/discount'
+import { Price } from '../../api-models/price'
 import { Product } from '../../api-models/product'
 import { CartProduct } from '../ui/cart-product'
 
@@ -6,9 +8,9 @@ export interface Cart {
     count: number
     items: Product[]
     displayItems: CartProduct[]
-    subTotal: number
-    total: number
-    discounts: string[]
+    subTotal: Price
+    total: Price
+    discounts: Discount[]
 }
 
 export interface CartDocument extends Cart, Document { }
